+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/myCommon.h
@@ -2,6 +2,7 @@
 #define __MYCOMMON_H__
 
 #include "Communication.h"
+#include "receiverBoard.h"
 
   void getMyMsg();
